[lab2]: finish implementation of initrdfs and fix some bug

+++ lab2/peripherals/mailbox.c
@@ -56,7 +56,7 @@ static uint32_t MBox_process(uint8_t* tags, size_t tags_size){
     }
 
     // write to mailbox
-    MBox_write((uint32_t)MBox_buf, MBOX_CHANNEL_PROPERTY_TAGS);
+    MBox_write((uint32_t)(uint64_t)MBox_buf, MBOX_CHANNEL_PROPERTY_TAGS);
     
     // read from mailbox
     uint32_t result = MBox_read(MBOX_CHANNEL_PROPERTY_TAGS);
@@ -76,7 +76,7 @@ void MBox_get_board_revision(uint32_t* ret){
     ((uint32_t*)tag->values)[0] = 0;
     
     // process tags
-    MBox_buffer* mbuf = (MBox_buffer *)MBox_process(buf, sizeof(MBox_tag) + 4);
+    MBox_buffer* mbuf = (MBox_buffer *)(uint64_t)MBox_process(buf, sizeof(MBox_tag) + 4);
 
     // precoess response tags
     tag = (MBox_tag*)mbuf->buf;
@@ -100,7 +100,7 @@ void MBox_get_arm_memory(uint32_t* ret){
     ((uint32_t*)tag->values)[1] = 0;
     
     // process tags
-    MBox_buffer* mbuf = (MBox_buffer *)MBox_process(buf, sizeof(MBox_tag) + 8);
+    MBox_buffer* mbuf = (MBox_buffer *)(uint64_t)MBox_process(buf, sizeof(MBox_tag) + 8);
 
     // precoess response tags
     tag = (MBox_tag*)mbuf->buf;
